Phase 1: cross-platform refactor for Windows/Linux/other support
Splits every macOS-specific call behind //go:build tags so future Windows
and Linux implementations can drop in without disturbing the macOS
codepath. No behavior change on macOS.

Per-OS dispatchers introduced:
- internal/cert: TrustCA per cert_{darwin,linux,windows,other}.go
- internal/daemon: process spawn (buildShellCommand/applySpawnAttrs/
  killProcessTree) per process_{darwin,linux,windows,other}.go;
  processAlive per _alive_{unix,windows}.go; lsof/ps/SIGTERM helpers
  per daemon_{unix,windows}.go; port discovery per
  port_discover_{unix,windows}.go
- cmd: setup, daemon lifecycle (tryPlatformDaemonStart/Stop, forkDaemon,
  cliProcessAlive, signalDaemonStop), install destination, and browser
  opener all split per-OS

macOS keeps \$SHELL -lic exactly — load-bearing for nvm/rbenv/pyenv
managed routes. Linux uses \$SHELL -c. Windows uses %COMSPEC% /C as a
Phase 1 stub. Job Object supervision and embedded DNS/netsh portproxy
land in Phase 2.

CI matrix now runs build/vet/test on macos-latest, ubuntu-latest, and
windows-latest. Tests that depend on unix shell commands or syscall
fields not present on Windows (api_test.go, monitor_test.go,
process_test.go) are tagged !windows; testServer() moved to a shared
test_helpers_test.go so cross-platform test files still compile.
cert_test.go skips unix permission-bit assertions on Windows;
daemon_test.go sets both HOME and USERPROFILE for portable home-dir
override.

Adds setup.ps1 — Windows install bootstrap that mirrors setup.sh:
installs Go via winget if missing, builds vibe.exe, copies it into
%LOCALAPPDATA%\Programs\vibe\, adds to user PATH, and starts the daemon.
README gains Windows + Linux install sections.
