+++ cmd/install_linux.go
@@ -1,0 +1,7 @@
+//go:build linux
+
+package cmd
+
+func installDestination() string {
+	return "/usr/local/bin/vibe"
+}
