+++ cmd/install_other.go
@@ -1,0 +1,7 @@
+//go:build !darwin && !linux && !windows
+
+package cmd
+
+func installDestination() string {
+	return "/usr/local/bin/vibe"
+}
