+++ cmd/open_darwin.go
@@ -1,0 +1,10 @@
+//go:build darwin
+
+package cmd
+
+import "os/exec"
+
+// openURL opens url in the default browser via the macOS `open` command.
+func openURL(url string) error {
+	return exec.Command("open", url).Run()
+}
