+++ internal/cert/cert.go
@@ -11,7 +11,6 @@ import (
 	"fmt"
 	"math/big"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"time"
 )
@@ -184,28 +183,11 @@ func LoadTLSConfig(certFile, keyFile string) (*tls.Config, error) {
 	}, nil
 }
 
-// TrustCA installs the CA certificate into the macOS System Keychain.
-// Requires root privileges.
-func TrustCA(certsDir string) error {
-	certPath := filepath.Join(certsDir, "ca.pem")
-
-	// Check if already trusted
-	out, err := exec.Command("security", "find-certificate", "-c", "local.vibe CA", "/Library/Keychains/System.keychain").CombinedOutput()
-	if err == nil && len(out) > 0 {
-		// Already trusted — remove old one first so we can update
-		_ = exec.Command("security", "delete-certificate", "-c", "local.vibe CA", "/Library/Keychains/System.keychain").Run()
-	}
-
-	cmd := exec.Command("security", "add-trusted-cert",
-		"-d", "-r", "trustRoot",
-		"-k", "/Library/Keychains/System.keychain",
-		certPath,
-	)
-	if output, err := cmd.CombinedOutput(); err != nil {
-		return fmt.Errorf("trust CA: %w — %s", err, string(output))
-	}
-	return nil
-}
+// TrustCA installs the CA certificate into the OS-level trust store.
+// Implementation lives in cert_<goos>.go. Requires elevated privileges.
+//
+// Note: Linux and Windows implementations are stubs in Phase 1 of Windows
+// support; only macOS is fully wired. See FUTURE_PLAN.md for status.
 
 // --- helpers ---
